Fix post service location not being corrected after refactoring.
