+++ PhotoPlusFrontend/PhotoPlus/src/app/components/forum/post/post.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input, ViewChild, ElementRef, Output, EventEmitter } from '@angular/core';
 import { Post } from 'src/app/models/post/post';
-import { PostService } from 'src/app/services/post.service';
+import { PostService } from 'src/app/services/post/post.service';
 import { LoginService } from 'src/app/services/login/login.service';
 import { UserService } from 'src/app/services/user/user.service';
 import { BehaviorSubject } from 'rxjs';
